refactor(cache): restructure the cache entry around fs and env fingerprint records
The cache entry conflated three records: the listed inputs' pre-run content
hashes (globbed_inputs), the fspy-discovered input fingerprints, and the
tracked env state — the first two scattered across the entry and the post-run
fingerprint, with the change-kind vocabulary living in the cache module (a
module cycle with the fingerprint module).

Regroup them by what they mean: InputFingerprints (everything the run read —
snapshot + discovered halves, one opaque serialized value) and
TrackedEnvFingerprints (everything runner-aware tools observed in the env),
each owning its validation. Lookup and update control flow is unchanged.

The entry layout changes, so CACHE_SCHEMA_VERSION bumps 18 -> 19; existing
caches are ignored once via the versioned cache directory.
